+++ src/routes/_auth/login/index.svelte
@@ -2,8 +2,8 @@
 	import { navigate } from "sv-router/generated";
     import { signInWithEmail } from "@/lib/auth/service";
     import { Eye, EyeOff, LogIn, UserPlus } from "@lucide/svelte";
-    import { getPlatform } from "@/lib/shared/platform";
     import { authentication } from "@/lib/auth/store.svelte";
+    import { app } from "@/lib/app/main";
 
     // Inputs for logging into the website, changes in the input changes the values.
     let email: string = $state("");
@@ -42,7 +42,7 @@
     <section class="
         p-6 flex flex-col
         justify-between
-        {(getPlatform() === "android" ) ? "pb-12" : ""}
+        {(app.platform === "android" ) ? "pb-12" : ""}
         sm:flex-1 sm:justify-center sm:h-auto
     ">
         <div class="